Add Auth0 login and logout buttons, and refactor AddButtons component
