+++ Breeze.Web/src/components/auth0/Login.tsx
@@ -1,6 +1,9 @@
 import { useAuth0 } from '@auth0/auth0-react'
 import { BreezeButton } from '../shared/BreezeButton'
 
+/**
+ * Simple login button from Auth0 docs.
+ */
 const LoginButton = () => {
 	const { loginWithRedirect } = useAuth0()
